GH-36: Move CompoundAST to Tag.rs and add comments to uncommented code
